+++ src/services/show-error.ts
@@ -1,13 +1,9 @@
-import {Alert} from 'react-native';
+import {showAlertDialog} from './dialogs';
 
 export function showError(title: string, message: string, dismissLabel?: string, onError?: (title: string, message: string) => void) {
     if (onError) {
         onError(title, message);
     } else {
-        Alert.alert(
-            title,
-            message,
-            [{text: dismissLabel || 'Dismiss'}]
-        );
+        showAlertDialog(title, message, dismissLabel || 'Dismiss');
     }
 }